refactoring entity packages
